Allow Northwoord drivers to be built if TAB driver is disabled (but without write support for NWT_GRD)
