+++ frmts/CMakeLists.txt
@@ -85,9 +85,7 @@ gdal_optional_format(blx "Magellan BLX Topo File Format")
 gdal_optional_format(msgn "Meteosat Second Generation (MSG) Native Archive Format (.nat)")
 gdal_optional_format(til "EarthWatch .TIL Driver")
 gdal_optional_format(r "R Object Data Store")
-if (NOT OGR_ENABLE_DRIVER_TAB_PLUGIN)
-  gdal_dependent_format(northwood "NWT_GRD/NWT_GRC -- Northwood/Vertical Mapper File Format" "OGR_ENABLE_DRIVER_TAB")
-endif()
+gdal_optional_format(northwood "NWT_GRD/NWT_GRC -- Northwood/Vertical Mapper File Format")
 gdal_optional_format(saga "SAGA GIS Binary Driver")
 gdal_optional_format(xyz "ASCII Gridded XYZ")
 include(avif/driver_declaration.cmake)
