+++ ogr/ogrmitabspatialref.cpp
@@ -1,36 +1,14 @@
-/**********************************************************************
- *
- * Name:     mitab_spatialref.cpp
- * Project:  MapInfo TAB Read/Write library
- * Language: C++
- * Purpose:  Implementation of the SpatialRef stuff in the TABFile class.
- * Author:   Frank Warmerdam, warmerdam@pobox.com
- *
- **********************************************************************
- * Copyright (c) 1999-2001, Frank Warmerdam
- * Copyright (c) 2014, Even Rouault <even.rouault at spatialys.com>
- *
- * Permission is hereby granted, free of charge, to any person obtaining a
- * copy of this software and associated documentation files (the "Software"),
- * to deal in the Software without restriction, including without limitation
- * the rights to use, copy, modify, merge, publish, distribute, sublicense,
- * and/or sell copies of the Software, and to permit persons to whom the
- * Software is furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included
- * in all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.  IN NO EVENT SHALL
- * THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
- * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
- * DEALINGS IN THE SOFTWARE.
- **********************************************************************/
+// SPDX-License-Identifier: MIT
+// Copyright 1999-2003, Daniel Morissette
+// Copyright (c) 1999-2001, Frank Warmerdam
+// Implementation translation between MIF CoordSys format, and
+// and OGRSpatialRef format.
+
+/*! @cond Doxygen_Suppress */
+
+#include "ogrmitabspatialref.h"
 
 #include "cpl_port.h"
-#include "mitab.h"
 
 #include <cmath>
 #include <cstddef>
@@ -41,9 +19,94 @@
 #include "cpl_conv.h"
 #include "cpl_error.h"
 #include "cpl_string.h"
-#include "mitab_priv.h"
 #include "ogr_spatialref.h"
 #include "ogr_srs_api.h"
+
+typedef struct
+{
+    int nDatumEPSGCode;
+    int nMapInfoDatumID;
+    const char *pszOGCDatumName;
+    int nEllipsoid;
+    double dfShiftX;
+    double dfShiftY;
+    double dfShiftZ;
+    double dfDatumParm0; /* RotX */
+    double dfDatumParm1; /* RotY */
+    double dfDatumParm2; /* RotZ */
+    double dfDatumParm3; /* Scale Factor */
+    double dfDatumParm4; /* Prime Meridian */
+} MapInfoDatumInfo;
+
+typedef struct
+{
+    int nMapInfoId;
+    const char *pszMapinfoName;
+    double dfA;             /* semi major axis in meters */
+    double dfInvFlattening; /* Inverse flattening */
+} MapInfoSpheroidInfo;
+
+/**********************************************************************
+ * MapInfo Units string to numeric ID conversion
+ **********************************************************************/
+typedef struct
+{
+    int nUnitId;
+    const char *pszAbbrev;
+} MapInfoUnitsInfo;
+
+static const MapInfoUnitsInfo gasUnitsList[] = {
+    {0, "mi"},        {1, "km"},          {2, "in"},  {3, "ft"},
+    {4, "yd"},        {5, "mm"},          {6, "cm"},  {7, "m"},
+    {8, "survey ft"}, {8, "survey foot"},  // alternate
+    {13, nullptr},    {9, "nmi"},         {30, "li"}, {31, "ch"},
+    {32, "rd"},       {-1, nullptr}};
+
+/**********************************************************************
+ *                       TABUnitIdToString()
+ *
+ * Return the MIF units name for specified units id.
+ * Return "" if no match found.
+ *
+ * The returned string should not be freed by the caller.
+ **********************************************************************/
+static const char *TABUnitIdToString(int nId)
+{
+    const MapInfoUnitsInfo *psList = gasUnitsList;
+
+    while (psList->nUnitId != -1)
+    {
+        if (psList->nUnitId == nId)
+            return psList->pszAbbrev;
+        psList++;
+    }
+
+    return "";
+}
+
+/**********************************************************************
+ *                       TABUnitIdFromString()
+ *
+ * Return the units ID for specified MIF units name
+ *
+ * Returns -1 if no match found.
+ **********************************************************************/
+static int TABUnitIdFromString(const char *pszName)
+{
+    if (pszName == nullptr)
+        return 13;
+
+    const MapInfoUnitsInfo *psList = gasUnitsList;
+
+    while (psList->nUnitId != -1)
+    {
+        if (psList->pszAbbrev != nullptr && EQUAL(psList->pszAbbrev, pszName))
+            return psList->nUnitId;
+        psList++;
+    }
+
+    return -1;
+}
 
 /* -------------------------------------------------------------------- */
 /*      This table was automatically generated by doing translations    */
@@ -692,54 +755,6 @@ static const MapInfoLCCSRS asMapInfoLCCSRSList[] = {
 };
 
 /**********************************************************************
- *                   TABFile::GetSpatialRef()
- *
- * Returns a reference to an OGRSpatialReference for this dataset.
- * If the projection parameters have not been parsed yet, then we will
- * parse them before returning.
- *
- * The returned object is owned and maintained by this TABFile and
- * should not be modified or freed by the caller.
- *
- * Returns NULL if the SpatialRef cannot be accessed.
- **********************************************************************/
-OGRSpatialReference *TABFile::GetSpatialRef()
-{
-    if (m_poMAPFile == nullptr)
-    {
-        CPLError(CE_Failure, CPLE_AssertionFailed,
-                 "GetSpatialRef() failed: file has not been opened yet.");
-        return nullptr;
-    }
-
-    if (GetGeomType() == wkbNone)
-        return nullptr;
-
-    /*-----------------------------------------------------------------
-     * If projection params have already been processed, just use them.
-     *----------------------------------------------------------------*/
-    if (m_poSpatialRef != nullptr)
-        return m_poSpatialRef;
-
-    /*-----------------------------------------------------------------
-     * Fetch the parameters from the header.
-     *----------------------------------------------------------------*/
-    TABProjInfo sTABProj;
-
-    TABMAPHeaderBlock *poHeader = nullptr;
-    if ((poHeader = m_poMAPFile->GetHeaderBlock()) == nullptr ||
-        poHeader->GetProjInfo(&sTABProj) != 0)
-    {
-        CPLError(CE_Failure, CPLE_FileIO,
-                 "GetSpatialRef() failed reading projection parameters.");
-        return nullptr;
-    }
-
-    m_poSpatialRef = GetSpatialRefFromTABProj(sTABProj);
-    return m_poSpatialRef;
-}
-
-/**********************************************************************
  *                   TABFile::GetSpatialRefFromTABProj()
  **********************************************************************/
 
@@ -750,7 +765,7 @@ static bool TAB_EQUAL(double a, double b)
 }
 
 OGRSpatialReference *
-TABFile::GetSpatialRefFromTABProj(const TABProjInfo &sTABProj)
+TABFileGetSpatialRefFromTABProj(const TABProjInfo &sTABProj)
 {
     /*-----------------------------------------------------------------
      * Get the units name, and translation factor.
@@ -1432,65 +1447,6 @@ TABFile::GetSpatialRefFromTABProj(const TABProjInfo &sTABProj)
     return poSpatialRef;
 }
 
-/**********************************************************************
- *                   TABFile::SetSpatialRef()
- *
- * Set the OGRSpatialReference for this dataset.
- * A reference to the OGRSpatialReference will be kept, and it will also
- * be converted into a TABProjInfo to be stored in the .MAP header.
- *
- * Returns 0 on success, and -1 on error.
- **********************************************************************/
-int TABFile::SetSpatialRef(OGRSpatialReference *poSpatialRef)
-{
-    if (m_eAccessMode != TABWrite)
-    {
-        CPLError(CE_Failure, CPLE_NotSupported,
-                 "SetSpatialRef() can be used only with Write access.");
-        return -1;
-    }
-
-    if (m_poMAPFile == nullptr)
-    {
-        CPLError(CE_Failure, CPLE_AssertionFailed,
-                 "SetSpatialRef() failed: file has not been opened yet.");
-        return -1;
-    }
-
-    if (poSpatialRef == nullptr)
-    {
-        CPLError(CE_Failure, CPLE_AssertionFailed,
-                 "SetSpatialRef() failed: Called with NULL poSpatialRef.");
-        return -1;
-    }
-
-    /*-----------------------------------------------------------------
-     * Keep a copy of the OGRSpatialReference...
-     * Note: we have to take the reference count into account...
-     *----------------------------------------------------------------*/
-    if (m_poSpatialRef && m_poSpatialRef->Dereference() == 0)
-        delete m_poSpatialRef;
-
-    m_poSpatialRef = poSpatialRef->Clone();
-
-    TABProjInfo sTABProj;
-    int nParamCount = 0;
-    GetTABProjFromSpatialRef(poSpatialRef, sTABProj, nParamCount);
-
-    /*-----------------------------------------------------------------
-     * Set the new parameters in the .MAP header.
-     * This will also trigger lookup of default bounds for the projection.
-     *----------------------------------------------------------------*/
-    if (SetProjInfo(&sTABProj) != 0)
-    {
-        CPLError(CE_Failure, CPLE_FileIO,
-                 "SetSpatialRef() failed setting projection parameters.");
-        return -1;
-    }
-
-    return 0;
-}
-
 static int MITABGetCustomDatum(const OGRSpatialReference *poSpatialRef,
                                TABProjInfo &sTABProj)
 {
@@ -1549,8 +1505,8 @@ static int MITABGetCustomDatum(const OGRSpatialReference *poSpatialRef,
     return TRUE;
 }
 
-int TABFile::GetTABProjFromSpatialRef(const OGRSpatialReference *poSpatialRef,
-                                      TABProjInfo &sTABProj, int &nParamCount)
+int TABFileGetTABProjFromSpatialRef(const OGRSpatialReference *poSpatialRef,
+                                    TABProjInfo &sTABProj, int &nParamCount)
 {
     /*-----------------------------------------------------------------
      * Initialize TABProjInfo
@@ -2318,3 +2274,349 @@ int TABFile::GetTABProjFromSpatialRef(const OGRSpatialReference *poSpatialRef,
 
     return 0;
 }
+
+extern const MapInfoDatumInfo asDatumInfoList[];
+extern const MapInfoSpheroidInfo asSpheroidInfoList[];
+
+/************************************************************************/
+/*                      MITABCoordSys2SpatialRef()                      */
+/*                                                                      */
+/*      Convert a MIF COORDSYS string into a new OGRSpatialReference    */
+/*      object.                                                         */
+/************************************************************************/
+
+OGRSpatialReference *MITABCoordSys2SpatialRef(const char *pszCoordSys)
+
+{
+    TABProjInfo sTABProj;
+    if (MITABCoordSys2TABProjInfo(pszCoordSys, &sTABProj) < 0)
+        return nullptr;
+    OGRSpatialReference *poSR = TABFileGetSpatialRefFromTABProj(sTABProj);
+
+    // Report on translation.
+    char *pszWKT = nullptr;
+
+    poSR->exportToWkt(&pszWKT);
+    if (pszWKT != nullptr)
+    {
+        CPLDebug("MITAB", "This CoordSys value:\n%s\nwas translated to:\n%s",
+                 pszCoordSys, pszWKT);
+        CPLFree(pszWKT);
+    }
+
+    return poSR;
+}
+
+/************************************************************************/
+/*                      MITABSpatialRef2CoordSys()                      */
+/*                                                                      */
+/*      Converts a OGRSpatialReference object into a MIF COORDSYS       */
+/*      string.                                                         */
+/*                                                                      */
+/*      The function returns a newly allocated string that should be    */
+/*      CPLFree()'d by the caller.                                      */
+/************************************************************************/
+
+char *MITABSpatialRef2CoordSys(const OGRSpatialReference *poSR)
+
+{
+    if (poSR == nullptr)
+        return nullptr;
+
+    TABProjInfo sTABProj;
+    int nParamCount = 0;
+    TABFileGetTABProjFromSpatialRef(poSR, sTABProj, nParamCount);
+
+    // Do coordsys lookup.
+    double dXMin = 0.0;
+    double dYMin = 0.0;
+    double dXMax = 0.0;
+    double dYMax = 0.0;
+    bool bHasBounds = false;
+    if (sTABProj.nProjId > 1 &&
+        MITABLookupCoordSysBounds(&sTABProj, dXMin, dYMin, dXMax, dYMax, true))
+    {
+        bHasBounds = true;
+    }
+
+    // Translate the units.
+    const char *pszMIFUnits = TABUnitIdToString(sTABProj.nUnitsId);
+
+    // Build coordinate system definition.
+    CPLString osCoordSys;
+
+    if (sTABProj.nProjId != 0)
+    {
+        osCoordSys.Printf("Earth Projection %d", sTABProj.nProjId);
+    }
+    else
+    {
+        osCoordSys.Printf("NonEarth Units");
+    }
+
+    // Append Datum.
+    if (sTABProj.nProjId != 0)
+    {
+        osCoordSys += CPLSPrintf(", %d", sTABProj.nDatumId);
+
+        if (sTABProj.nDatumId == 999 || sTABProj.nDatumId == 9999)
+        {
+            osCoordSys +=
+                CPLSPrintf(", %d, %.15g, %.15g, %.15g", sTABProj.nEllipsoidId,
+                           sTABProj.dDatumShiftX, sTABProj.dDatumShiftY,
+                           sTABProj.dDatumShiftZ);
+        }
+
+        if (sTABProj.nDatumId == 9999)
+        {
+            osCoordSys +=
+                CPLSPrintf(", %.15g, %.15g, %.15g, %.15g, %.15g",
+                           sTABProj.adDatumParams[0], sTABProj.adDatumParams[1],
+                           sTABProj.adDatumParams[2], sTABProj.adDatumParams[3],
+                           sTABProj.adDatumParams[4]);
+        }
+    }
+
+    // Append units.
+    if (sTABProj.nProjId != 1 && pszMIFUnits != nullptr)
+    {
+        if (sTABProj.nProjId != 0)
+            osCoordSys += ",";
+
+        osCoordSys += CPLSPrintf(" \"%s\"", pszMIFUnits);
+    }
+
+    // Append Projection Params.
+    for (int iParam = 0; iParam < nParamCount; iParam++)
+        osCoordSys += CPLSPrintf(", %.15g", sTABProj.adProjParams[iParam]);
+
+    // Append user bounds.
+    if (bHasBounds)
+    {
+        if (fabs(dXMin - floor(dXMin + 0.5)) < 1e-8 &&
+            fabs(dYMin - floor(dYMin + 0.5)) < 1e-8 &&
+            fabs(dXMax - floor(dXMax + 0.5)) < 1e-8 &&
+            fabs(dYMax - floor(dYMax + 0.5)) < 1e-8)
+        {
+            osCoordSys +=
+                CPLSPrintf(" Bounds (%d, %d) (%d, %d)", static_cast<int>(dXMin),
+                           static_cast<int>(dYMin), static_cast<int>(dXMax),
+                           static_cast<int>(dYMax));
+        }
+        else
+        {
+            osCoordSys += CPLSPrintf(" Bounds (%f, %f) (%f, %f)", dXMin, dYMin,
+                                     dXMax, dYMax);
+        }
+    }
+
+    // Report on translation.
+    char *pszWKT = nullptr;
+
+    poSR->exportToWkt(&pszWKT);
+    if (pszWKT != nullptr)
+    {
+        CPLDebug("MITAB", "This WKT Projection:\n%s\n\ntranslates to:\n%s",
+                 pszWKT, osCoordSys.c_str());
+        CPLFree(pszWKT);
+    }
+
+    return CPLStrdup(osCoordSys.c_str());
+}
+
+/************************************************************************/
+/*                      MITABExtractCoordSysBounds                      */
+/*                                                                      */
+/* Return true if MIF coordsys string contains a BOUNDS parameter and   */
+/* Set x/y min/max values.                                              */
+/************************************************************************/
+
+bool MITABExtractCoordSysBounds(const char *pszCoordSys, double &dXMin,
+                                double &dYMin, double &dXMax, double &dYMax)
+
+{
+    if (pszCoordSys == nullptr)
+        return false;
+
+    char **papszFields =
+        CSLTokenizeStringComplex(pszCoordSys, " ,()", TRUE, FALSE);
+
+    int iBounds = CSLFindString(papszFields, "Bounds");
+
+    if (iBounds >= 0 && iBounds + 4 < CSLCount(papszFields))
+    {
+        dXMin = CPLAtof(papszFields[++iBounds]);
+        dYMin = CPLAtof(papszFields[++iBounds]);
+        dXMax = CPLAtof(papszFields[++iBounds]);
+        dYMax = CPLAtof(papszFields[++iBounds]);
+        CSLDestroy(papszFields);
+        return true;
+    }
+
+    CSLDestroy(papszFields);
+    return false;
+}
+
+/**********************************************************************
+ *                     MITABCoordSys2TABProjInfo()
+ *
+ * Convert a MIF COORDSYS string into a TABProjInfo structure.
+ *
+ * Returns 0 on success, -1 on error.
+ **********************************************************************/
+int MITABCoordSys2TABProjInfo(const char *pszCoordSys, TABProjInfo *psProj)
+
+{
+    // Set all fields to zero, equivalent of NonEarth Units "mi"
+    memset(psProj, 0, sizeof(TABProjInfo));
+
+    if (pszCoordSys == nullptr)
+        return -1;
+
+    // Parse the passed string into words.
+    while (*pszCoordSys == ' ')
+        pszCoordSys++;  // Eat leading spaces.
+    if (STARTS_WITH_CI(pszCoordSys, "CoordSys") && pszCoordSys[8] != '\0')
+        pszCoordSys += 9;
+
+    char **papszFields =
+        CSLTokenizeStringComplex(pszCoordSys, " ,", TRUE, FALSE);
+
+    // Clip off Bounds information.
+    int iBounds = CSLFindString(papszFields, "Bounds");
+
+    while (iBounds != -1 && papszFields[iBounds] != nullptr)
+    {
+        CPLFree(papszFields[iBounds]);
+        papszFields[iBounds] = nullptr;
+        iBounds++;
+    }
+
+    // Fetch the projection.
+    char **papszNextField = nullptr;
+
+    if (CSLCount(papszFields) >= 3 && EQUAL(papszFields[0], "Earth") &&
+        EQUAL(papszFields[1], "Projection"))
+    {
+        int nProjId = atoi(papszFields[2]);
+        if (nProjId >= 3000)
+            nProjId -= 3000;
+        else if (nProjId >= 2000)
+            nProjId -= 2000;
+        else if (nProjId >= 1000)
+            nProjId -= 1000;
+
+        psProj->nProjId = static_cast<GByte>(nProjId);
+        papszNextField = papszFields + 3;
+    }
+    else if (CSLCount(papszFields) >= 2 && EQUAL(papszFields[0], "NonEarth"))
+    {
+        // NonEarth Units "..." Bounds (x, y) (x, y)
+        psProj->nProjId = 0;
+        papszNextField = papszFields + 2;
+
+        if (papszNextField[0] != nullptr && EQUAL(papszNextField[0], "Units"))
+            papszNextField++;
+    }
+    else
+    {
+        // Invalid projection string ???
+        if (CSLCount(papszFields) > 0)
+            CPLError(CE_Warning, CPLE_IllegalArg,
+                     "Failed parsing CoordSys: '%s'", pszCoordSys);
+        CSLDestroy(papszFields);
+        return -1;
+    }
+
+    // Fetch the datum information.
+    int nDatum = 0;
+
+    if (psProj->nProjId != 0 && CSLCount(papszNextField) > 0)
+    {
+        nDatum = atoi(papszNextField[0]);
+        papszNextField++;
+    }
+
+    if ((nDatum == 999 || nDatum == 9999) && CSLCount(papszNextField) >= 4)
+    {
+        psProj->nEllipsoidId = static_cast<GByte>(atoi(papszNextField[0]));
+        psProj->dDatumShiftX = CPLAtof(papszNextField[1]);
+        psProj->dDatumShiftY = CPLAtof(papszNextField[2]);
+        psProj->dDatumShiftZ = CPLAtof(papszNextField[3]);
+        papszNextField += 4;
+
+        if (nDatum == 9999 && CSLCount(papszNextField) >= 5)
+        {
+            psProj->adDatumParams[0] = CPLAtof(papszNextField[0]);
+            psProj->adDatumParams[1] = CPLAtof(papszNextField[1]);
+            psProj->adDatumParams[2] = CPLAtof(papszNextField[2]);
+            psProj->adDatumParams[3] = CPLAtof(papszNextField[3]);
+            psProj->adDatumParams[4] = CPLAtof(papszNextField[4]);
+            papszNextField += 5;
+        }
+    }
+    else if (nDatum != 999 && nDatum != 9999)
+    {
+        // Find the datum, and collect its parameters if possible.
+        const MapInfoDatumInfo *psDatumInfo = nullptr;
+
+        int iDatum = 0;  // Used after for.
+        for (; asDatumInfoList[iDatum].nMapInfoDatumID != -1; iDatum++)
+        {
+            if (asDatumInfoList[iDatum].nMapInfoDatumID == nDatum)
+            {
+                psDatumInfo = asDatumInfoList + iDatum;
+                break;
+            }
+        }
+
+        if (asDatumInfoList[iDatum].nMapInfoDatumID == -1)
+        {
+            // Use WGS84.
+            psDatumInfo = asDatumInfoList + 0;
+        }
+
+        if (psDatumInfo != nullptr)
+        {
+            psProj->nEllipsoidId = static_cast<GByte>(psDatumInfo->nEllipsoid);
+            psProj->nDatumId =
+                static_cast<GInt16>(psDatumInfo->nMapInfoDatumID);
+            psProj->dDatumShiftX = psDatumInfo->dfShiftX;
+            psProj->dDatumShiftY = psDatumInfo->dfShiftY;
+            psProj->dDatumShiftZ = psDatumInfo->dfShiftZ;
+            psProj->adDatumParams[0] = psDatumInfo->dfDatumParm0;
+            psProj->adDatumParams[1] = psDatumInfo->dfDatumParm1;
+            psProj->adDatumParams[2] = psDatumInfo->dfDatumParm2;
+            psProj->adDatumParams[3] = psDatumInfo->dfDatumParm3;
+            psProj->adDatumParams[4] = psDatumInfo->dfDatumParm4;
+        }
+    }
+
+    // Fetch the units string.
+    if (CSLCount(papszNextField) > 0)
+    {
+        if (isdigit(static_cast<unsigned char>(papszNextField[0][0])))
+        {
+            psProj->nUnitsId = static_cast<GByte>(atoi(papszNextField[0]));
+        }
+        else
+        {
+            psProj->nUnitsId =
+                static_cast<GByte>(TABUnitIdFromString(papszNextField[0]));
+        }
+        papszNextField++;
+    }
+
+    // Finally the projection parameters.
+    for (int iParam = 0; iParam < 7 && CSLCount(papszNextField) > 0; iParam++)
+    {
+        psProj->adProjParams[iParam] = CPLAtof(papszNextField[0]);
+        papszNextField++;
+    }
+
+    CSLDestroy(papszFields);
+
+    return 0;
+}
+
+/*! @endcond */
