+++ ogr/ogrsf_frmts/mitab/mitab_priv.h
@@ -36,6 +36,7 @@
 #include "cpl_conv.h"
 #include "cpl_string.h"
 #include "ogr_feature.h"
+#include "ogrmitabspatialref.h"
 
 #include <set>
 
@@ -255,34 +256,6 @@ typedef struct TABMAPCoordSecHdr_t
     GInt32 nDataOffset;
     int nVertexOffset;
 } TABMAPCoordSecHdr;
-
-/*---------------------------------------------------------------------
- * TABProjInfo
- * struct used to store the projection parameters from the .MAP header
- *--------------------------------------------------------------------*/
-typedef struct TABProjInfo_t
-{
-    GByte nProjId;  // See MapInfo Ref. Manual, App. F and G
-    GByte nEllipsoidId;
-    GByte nUnitsId;
-    double adProjParams[7];  // params in same order as in .MIF COORDSYS
-
-    GInt16 nDatumId;      // Datum Id added in MapInfo 7.8+ (.map V500)
-    double dDatumShiftX;  // Before that, we had to always lookup datum
-    double dDatumShiftY;  // parameters to establish datum id
-    double dDatumShiftZ;
-    double adDatumParams[5];
-
-    // Affine parameters only in .map version 500 and up
-    GByte nAffineFlag;  // 0=No affine param, 1=Affine params
-    GByte nAffineUnits;
-    double dAffineParamA;  // Affine params
-    double dAffineParamB;
-    double dAffineParamC;
-    double dAffineParamD;
-    double dAffineParamE;
-    double dAffineParamF;
-} TABProjInfo;
 
 /*---------------------------------------------------------------------
  * TABPenDef - Pen definition information
