+++ ogr/ogrspatialreference.cpp
@@ -55,6 +55,7 @@
 #include "ogr_p.h"
 #include "ogr_proj_p.h"
 #include "ogr_srs_api.h"
+#include "ogrmitabspatialref.h"
 
 #include "proj.h"
 #include "proj_experimental.h"
@@ -11208,11 +11209,6 @@ OGRErr OSRSetAxes(OGRSpatialReferenceH hSRS, const char *pszTargetKey,
                                     eYAxisOrientation);
 }
 
-#ifdef HAVE_MITAB
-char CPL_DLL *MITABSpatialRef2CoordSys(const OGRSpatialReference *);
-OGRSpatialReference CPL_DLL *MITABCoordSys2SpatialRef(const char *);
-#endif
-
 /************************************************************************/
 /*                       OSRExportToMICoordSys()                        */
 /************************************************************************/
@@ -11251,22 +11247,14 @@ OGRErr OSRExportToMICoordSys(OGRSpatialReferenceH hSRS, char **ppszReturn)
  * OGRERR_UNSUPPORTED_OPERATION if MITAB library was not linked in.
  */
 
-OGRErr OGRSpatialReference::exportToMICoordSys(
-    [[maybe_unused]] char **ppszResult) const
+OGRErr OGRSpatialReference::exportToMICoordSys(char **ppszResult) const
 
 {
-#ifdef HAVE_MITAB
     *ppszResult = MITABSpatialRef2CoordSys(this);
     if (*ppszResult != nullptr && strlen(*ppszResult) > 0)
         return OGRERR_NONE;
 
     return OGRERR_FAILURE;
-#else
-    CPLError(CE_Failure, CPLE_NotSupported,
-             "MITAB not available, CoordSys support disabled.");
-
-    return OGRERR_UNSUPPORTED_OPERATION;
-#endif
 }
 
 /************************************************************************/
@@ -11306,11 +11294,9 @@ OGRErr OSRImportFromMICoordSys(OGRSpatialReferenceH hSRS,
  * OGRERR_UNSUPPORTED_OPERATION if MITAB library was not linked in.
  */
 
-OGRErr OGRSpatialReference::importFromMICoordSys(
-    [[maybe_unused]] const char *pszCoordSys)
+OGRErr OGRSpatialReference::importFromMICoordSys(const char *pszCoordSys)
 
 {
-#ifdef HAVE_MITAB
     OGRSpatialReference *poResult = MITABCoordSys2SpatialRef(pszCoordSys);
 
     if (poResult == nullptr)
@@ -11320,12 +11306,6 @@ OGRErr OGRSpatialReference::importFromMICoordSys(
     delete poResult;
 
     return OGRERR_NONE;
-#else
-    CPLError(CE_Failure, CPLE_NotSupported,
-             "MITAB not available, CoordSys support disabled.");
-
-    return OGRERR_UNSUPPORTED_OPERATION;
-#endif
 }
 
 /************************************************************************/
